+++ include/coro_rpc/coro_rpc/coro_connection.hpp
@@ -27,10 +27,10 @@
 #include <vector>
 
 #include "asio/buffer.hpp"
-#include "asio_util/asio_coro_util.hpp"
-#include "asio_util/asio_util.hpp"
 #include "async_simple/Executor.h"
 #include "async_simple/coro/SyncAwait.h"
+#include "coro_io/asio_coro_util.hpp"
+#include "coro_io/asio_util.hpp"
 #include "easylog/easylog.h"
 #ifdef UNIT_TEST_INJECT
 #include "inject_action.hpp"
