+++ include/coro_rpc/coro_rpc/default_config/coro_rpc_config.hpp
@@ -18,7 +18,7 @@
 #include <chrono>
 #include <thread>
 
-#include "asio_util/io_context_pool.hpp"
+#include "coro_io/io_context_pool.hpp"
 #include "coro_rpc/coro_rpc/context.hpp"
 #include "coro_rpc/coro_rpc/protocol/coro_rpc_protocol.hpp"
 #include "coro_rpc/coro_rpc_server.hpp"
