+++ src/coro_rpc/benchmark/api/rpc_functions.hpp
@@ -20,8 +20,8 @@
 #include "Rect.h"
 #include "ValidateRequest.h"
 #include "asio/associated_executor.hpp"
-#include "asio_util/io_context_pool.hpp"
 #include "async_simple/coro/Lazy.h"
+#include "coro_io/io_context_pool.hpp"
 #include "coro_rpc/rpc_context.hpp"
 
 inline asio_util::io_context_pool pool(std::thread::hardware_concurrency());
