+++ src/coro_rpc/examples/user_defined_rpc_protocol/rest_rpc/config/rest_rpc_protocol.hpp
@@ -17,7 +17,7 @@
 
 #include <variant>
 
-#include "asio_util/io_context_pool.hpp"
+#include "coro_io/io_context_pool.hpp"
 #include "coro_rpc/coro_rpc/router.hpp"
 #include "msgpack_protocol.hpp"
