+++ src/coro_rpc/tests/test_coro_rpc_client.cpp
@@ -25,7 +25,7 @@
 #include <variant>
 
 #include "asio/io_context.hpp"
-#include "asio_util/asio_coro_util.hpp"
+#include "coro_io/asio_coro_util.hpp"
 #include "doctest.h"
 #include "rpc_api.hpp"
 #include "struct_pack/struct_pack.hpp"
